moved source code to src folder
