+++ package.json
@@ -18,7 +18,7 @@
   "activationEvents": [
     "onLanguage:asp"
   ],
-  "main": "./extension.js",
+  "main": "./src/extension.js",
   "contributes": {
     "viewsContainers": {
       "panel": [
